Removing old tests. Moving new tests to proper location in repository. Added first Contract test.
