--- CorDapp/contracts/src/test/kotlin/com/dcm/state/DataRowStateTests.kt
@@ -1,4 +1,0 @@
-package net.corda.training.com.dcm.state
-
-class DataRowStateTests {
-}
